requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
pytest
httpx
